myte/templates/template-flask-simple/requirements.txt
alembic
blinker
click
colorama
Flask
Flask-Migrate
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2-binary
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug